clang10 again

+++ src/test/operators/join_test_runner.cpp
@@ -237,7 +237,8 @@ class JoinTestRunner : public BaseTestWithParam<JoinTestConfiguration> {
 
           auto indexed_chunk_ranges = std::set<ChunkRange>{};
           for (const auto indexed_share : indexed_segment_shares) {
-            indexed_chunk_ranges.insert(ChunkRange{0, static_cast<uint32_t>(std::ceil(indexed_share * chunk_count))});
+            indexed_chunk_ranges.insert(
+                ChunkRange{0, static_cast<uint32_t>(std::ceil(indexed_share * static_cast<float>(chunk_count)))});
           }
 
           for (const auto& indexed_chunk_range : indexed_chunk_ranges) {
